+++ src/api/projects/deployments/deployment.rs
@@ -9,30 +9,34 @@ use derive_builder::Builder;
 use crate::api::common::NameOrId;
 use crate::api::endpoint_prelude::*;
 
-/// Delete a webhook for a project.
+/// Get a deployment for a project.
 #[derive(Debug, Builder, Clone)]
-pub struct DeleteHook<'a> {
-    /// The project to delete a webhook from.
+pub struct Deployment<'a> {
+    /// The project to get a deployment from.
     #[builder(setter(into))]
     project: NameOrId<'a>,
-    /// The ID of the hook to delete.
-    hook_id: u64,
+    /// The ID of the deployment.
+    deployment_id: u64,
 }
 
-impl<'a> DeleteHook<'a> {
+impl<'a> Deployment<'a> {
     /// Create a builder for the endpoint.
-    pub fn builder() -> DeleteHookBuilder<'a> {
-        DeleteHookBuilder::default()
+    pub fn builder() -> DeploymentBuilder<'a> {
+        DeploymentBuilder::default()
     }
 }
 
-impl<'a> Endpoint for DeleteHook<'a> {
+impl<'a> Endpoint for Deployment<'a> {
     fn method(&self) -> Method {
-        Method::DELETE
+        Method::GET
     }
 
     fn endpoint(&self) -> Cow<'static, str> {
-        format!("projects/{}/hooks/{}", self.project, self.hook_id).into()
+        format!(
+            "projects/{}/deployments/{}",
+            self.project, self.deployment_id,
+        )
+        .into()
     }
 }
 
@@ -40,36 +44,36 @@ impl<'a> Endpoint for DeleteHook<'a> {
 mod tests {
     use http::Method;
 
-    use crate::api::projects::hooks::{DeleteHook, DeleteHookBuilderError};
+    use crate::api::projects::deployments::{Deployment, DeploymentBuilderError};
     use crate::api::{self, Query};
     use crate::test::client::{ExpectedUrl, SingleTestClient};
 
     #[test]
-    fn project_and_hook_id_are_necessary() {
-        let err = DeleteHook::builder().build().unwrap_err();
-        crate::test::assert_missing_field!(err, DeleteHookBuilderError, "project");
+    fn project_and_deployment_id_are_necessary() {
+        let err = Deployment::builder().build().unwrap_err();
+        crate::test::assert_missing_field!(err, DeploymentBuilderError, "project");
     }
 
     #[test]
     fn project_is_necessary() {
-        let err = DeleteHook::builder().hook_id(1).build().unwrap_err();
-        crate::test::assert_missing_field!(err, DeleteHookBuilderError, "project");
+        let err = Deployment::builder().deployment_id(1).build().unwrap_err();
+        crate::test::assert_missing_field!(err, DeploymentBuilderError, "project");
     }
 
     #[test]
-    fn hook_id_is_necessary() {
-        let err = DeleteHook::builder()
+    fn deployment_id_is_necessary() {
+        let err = Deployment::builder()
             .project("project")
             .build()
             .unwrap_err();
-        crate::test::assert_missing_field!(err, DeleteHookBuilderError, "hook_id");
+        crate::test::assert_missing_field!(err, DeploymentBuilderError, "deployment_id");
     }
 
     #[test]
-    fn project_and_hook_id_are_sufficient() {
-        DeleteHook::builder()
+    fn project_and_deployment_id_are_sufficient() {
+        Deployment::builder()
             .project("project")
-            .hook_id(1)
+            .deployment_id(1)
             .build()
             .unwrap();
     }
@@ -77,15 +81,15 @@ mod tests {
     #[test]
     fn endpoint() {
         let endpoint = ExpectedUrl::builder()
-            .method(Method::DELETE)
-            .endpoint("projects/simple%2Fproject/hooks/1")
+            .method(Method::GET)
+            .endpoint("projects/simple%2Fproject/deployments/1")
             .build()
             .unwrap();
         let client = SingleTestClient::new_raw(endpoint, "");
 
-        let endpoint = DeleteHook::builder()
+        let endpoint = Deployment::builder()
             .project("simple/project")
-            .hook_id(1)
+            .deployment_id(1)
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();