+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Additions
 
   * Add `api::projects::DeleteProject` endpoint
+  * Add `api::projects::deployments::CreateDeployment` endpoint
+  * Add `api::projects::deployments::DeleteDeployment` endpoint
+  * Add `api::projects::deployments::Deployment` endpoint
+  * Add `api::projects::deployments::Deployments` endpoint
+  * Add `api::projects::deployments::EditDeployment` endpoint
 
 # v0.1700.1